refactored files to organize page routing
