+++ src/pages/IssueDebit.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../App.css';
+import '../pages/App.css';
 import Dropdown from 'react-dropdown';
 import 'react-dropdown/style.css';
